Encapsulate `PercentDrivenInteractiveTransition` as an abstract class for interactive animators
