+++ IBAnimatable/ScreenEdgePanInteractiveAnimator.swift
@@ -1,0 +1,10 @@
+//
+//  Created by Jake Lin on 4/5/16.
+//  Copyright © 2016 Jake Lin. All rights reserved.
+//
+
+import UIKit
+
+public class ScreenEdgePanInteractiveAnimator: UIPercentDrivenInteractiveTransition {
+
+}
